refactor(cli): deduplicate shared helpers into utils
Consolidate helpers that were defined identically in two places:

- formatDuration (create/bin.ts, migration/bin.ts) -> utils/terminal.ts
- findGitRoot (create/bin.ts, migration/migrator/git-hooks.ts) -> utils/git.ts
- cancelAndExit (create/prompts.ts duplicate removed) -> utils/prompts.ts
- getScopeFromPackageName (create/templates/monorepo.ts duplicate removed) -> utils/package.ts

parseArgs in create/bin.ts and migration/bin.ts share only a name and
parse different flags, so they are left as is. isPlainObject has a single
definition within the cli package.
